Added quick & dirty tool to merge soundfonts
The idea is to have the semantics of a new file while keeping the syntax as
close to the previous file as possible, so that the git diff isn't swamped
by too many irrelevant modifications.  This tool doesn't re-order generators
to conform with the SF2 specs yet, which makes it a hack for limited
applications.  Should be polished if scope is to increase.